rewrote stuff to use a better Parameter class
moved all of the necessary logic in a FunctionReference into Parameter/Function, so FunctionReference is not needed except to generate a Function

since this is a more advanced model the internal state has stronger requirements, this will break the program on some inputs.
should be changed by the next commit, since that will overhaul ParamSize.